Remove old code and final renames
